feat(mobile-ui): Expose UI module in sidebar
This just exposes the UI (previously titled Responsiveness) module in
the sidebar. I've changed the page filters to use our standard ones and
done a small refactor to share the same landing page code as the app
starts module.

Next will be modifying the view content to show the right charts at the
top and screens table content.
